+++ lib/provider/camera_mode.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:jigsaw_hints/camera_screen.dart';
+import 'package:jigsaw_hints/pages/camera_screen.dart';
 
 class CameraModeProvider extends ChangeNotifier {
-  /// Internal, private state of the notifier.
   CameraMode _mode = CameraMode.piece;
 
   CameraMode get mode => _mode;
@@ -10,7 +9,6 @@ class CameraModeProvider extends ChangeNotifier {
   set mode(CameraMode mode) {
     if (mode != _mode) {
       _mode = mode;
-      // This call tells the widgets that are listening to this model to rebuild.
       notifyListeners();
     }
   }
