+++ lib/settings/appearance.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:jigsaw_hints/app_bar.dart';
-import 'package:jigsaw_hints/constants.dart';
-import 'package:jigsaw_hints/main.dart';
+import 'package:jigsaw_hints/ui/app_bar.dart';
+import 'package:jigsaw_hints/utils/constants.dart';
+import 'package:jigsaw_hints/main/main.dart';
 import 'package:jigsaw_hints/settings/default_settings.dart';
 import 'package:jigsaw_hints/settings/shared_prefs.dart';
 import 'package:provider/provider.dart';
