+++ lib/settings/general.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:jigsaw_hints/app_bar.dart';
-import 'package:jigsaw_hints/constants.dart';
+import 'package:jigsaw_hints/ui/app_bar.dart';
+import 'package:jigsaw_hints/utils/constants.dart';
 import 'package:jigsaw_hints/settings/default_settings.dart';
 import 'package:jigsaw_hints/settings/input_dialog.dart';
 import 'package:jigsaw_hints/settings/shared_prefs.dart';
