+++ lib/settings/input_dialog.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:jigsaw_hints/constants.dart';
+import 'package:jigsaw_hints/utils/constants.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 StatefulBuilder inputDialogText(
