+++ lib/ui/app_bar.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'constants.dart';
+import '../utils/constants.dart';
 
 class JigsawAppBar extends StatelessWidget with PreferredSizeWidget {
   final String title;
